Switched to float64

+++ bmc/fid/engine.py
@@ -21,7 +21,7 @@ from bmc.utils.webhook import DiscordNotifier
 class BMCSim(BMCTool):
     def __init__(self, adc_time: float, params: Params, seq_file: str | Path, z_positions: torch.Tensor, n_backlog: str | int, verbose: bool = True, webhook: bool = False, **kwargs) -> None:
         super().__init__(params, seq_file, verbose, **kwargs)
-        self.z_positions = z_positions.to(GLOBAL_DEVICE, torch.float32)  # Torch-Tensor
+        self.z_positions = z_positions.to(GLOBAL_DEVICE, torch.float64)  # Torch-Tensor
         self.n_isochromats = len(self.z_positions)
         self.bm_solver = BlochMcConnellSolver(params=self.params, n_offsets=self.n_offsets, z_positions=self.z_positions)
         
@@ -39,11 +39,11 @@ class BMCSim(BMCTool):
         else:
             self.n_measure = self.n_offsets
 
-        self.m_out = torch.zeros(self.n_isochromats, self.m_init.shape[0], self.n_measure, dtype=torch.float32, device=GLOBAL_DEVICE)
-        self.m_out[:, :, 0] = torch.tensor(self.m_init, dtype=torch.float32, device=GLOBAL_DEVICE).unsqueeze(0)
+        self.m_out = torch.zeros(self.n_isochromats, self.m_init.shape[0], self.n_measure, dtype=torch.float64, device=GLOBAL_DEVICE)
+        self.m_out[:, :, 0] = torch.tensor(self.m_init, dtype=torch.float64, device=GLOBAL_DEVICE).unsqueeze(0)
 
         self.dt_adc = self.adc_time / self.params.options["max_pulse_samples"]
-        self.t = torch.tensor([0], dtype=torch.float32, device=GLOBAL_DEVICE)
+        self.t = torch.tensor([0], dtype=torch.float64, device=GLOBAL_DEVICE)
         self.total_vec = None
         self.events = []
 
@@ -82,7 +82,7 @@ class BMCSim(BMCTool):
             self.events.append(f'adc at {start_time.item():.4f}s')
 
             new_time_array = start_time + torch.arange(1, self.params.options["max_pulse_samples"] + 1, 
-                                                    dtype=torch.float32, device=GLOBAL_DEVICE) * self.dt_adc
+                                                    dtype=torch.float64, device=GLOBAL_DEVICE) * self.dt_adc
             local_t = torch.cat((self.t, new_time_array))
             adc_outputs = []
             
@@ -104,7 +104,7 @@ class BMCSim(BMCTool):
             if counter <= self.n_backlog:
                 start_time = self.t[-1]
                 self.events.append(f'rf at {start_time.item():.4f}s')
-                time_array = start_time + torch.arange(1, amp_.numel() + 1, dtype=torch.float32, device=GLOBAL_DEVICE) * dtp_
+                time_array = start_time + torch.arange(1, amp_.numel() + 1, dtype=torch.float64, device=GLOBAL_DEVICE) * dtp_
                 self.t = torch.cat((self.t, time_array))
             for i in range(amp_.numel()):
                 self.bm_solver.update_matrix(
@@ -122,7 +122,7 @@ class BMCSim(BMCTool):
                 mag = self.bm_solver.solve_equation(mag=mag, dtp=delay_after_pulse)
                 if counter <= self.n_backlog:
                     start_time = self.t[-1]
-                    time_array = start_time + torch.arange(1, 2, dtype=torch.float32, device=GLOBAL_DEVICE) * delay_after_pulse
+                    time_array = start_time + torch.arange(1, 2, dtype=torch.float64, device=GLOBAL_DEVICE) * delay_after_pulse
                     self.t = torch.cat((self.t, time_array))
                     self.m_out[:, :, current_adc] = mag.squeeze()
                     current_adc += 1
@@ -138,7 +138,7 @@ class BMCSim(BMCTool):
             if counter <= self.n_backlog:
                 start_time = self.t[-1]
                 self.events.append(f'gz at {start_time.item():.4f}s')
-                time_array = start_time + torch.arange(1, amp_.numel() + 1, dtype=torch.float32, device=GLOBAL_DEVICE) * dtp_
+                time_array = start_time + torch.arange(1, amp_.numel() + 1, dtype=torch.float64, device=GLOBAL_DEVICE) * dtp_
                 self.t = torch.cat((self.t, time_array))
             for i in range(amp_.numel()):
                 self.bm_solver.update_matrix(0, 0, 0, grad_amp=amp_[i])
@@ -152,7 +152,7 @@ class BMCSim(BMCTool):
                 mag = self.bm_solver.solve_equation(mag=mag, dtp=delay_after_grad)
                 if counter <= self.n_backlog:
                     start_time = self.t[-1]
-                    time_array = start_time + torch.arange(1, 2, dtype=torch.float32, device=GLOBAL_DEVICE) * delay_after_grad
+                    time_array = start_time + torch.arange(1, 2, dtype=torch.float64, device=GLOBAL_DEVICE) * delay_after_grad
                     self.t = torch.cat((self.t, time_array))
                     self.m_out[:, :, current_adc] = mag.squeeze()
                     current_adc += 1
@@ -165,7 +165,7 @@ class BMCSim(BMCTool):
             if counter <= self.n_backlog:
                 start_time = self.t[-1]
                 self.events.append(f'rf at {start_time.item():.4f}s')
-                time_array = start_time + torch.arange(1, sample_factor_delay + 1, dtype=torch.float32, device=GLOBAL_DEVICE) * dt_delay
+                time_array = start_time + torch.arange(1, sample_factor_delay + 1, dtype=torch.float64, device=GLOBAL_DEVICE) * dt_delay
                 self.t = torch.cat((self.t, time_array))
 
                 for step in range(len(time_array)):
@@ -194,7 +194,7 @@ class BMCSim(BMCTool):
 
         mag = torch.tensor(
             self.m_init[np.newaxis, np.newaxis, :, np.newaxis], 
-            dtype=torch.float32,
+            dtype=torch.float64,
             device=GLOBAL_DEVICE
         )
